backend/requirements.txt
APScheduler
bcrypt
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
Flask
Flask-APScheduler
flask-cors
Flask-JWT-Extended
Flask-SQLAlchemy
greenlet
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
packaging
psycopg2-binary
pycparser
PyJWT
PyMySQL
python-dateutil
python-dotenv
requests
six
SQLAlchemy
typing_extensions
tzdata
tzlocal
urllib3
Werkzeug
email-validator
boto3
prometheus-flask-exporter